server: check for exceeding maximum socket path length
At this time, Coverity did not have it flagged. But it is another
case like CID 43509.

+++ provider/server/ECSoapServerConnection.cpp
@@ -105,8 +105,11 @@ static int create_pipe_socket(const char *unix_socket, ECConfig *lpConfig,
 	
 	memset(&saddr,0,sizeof(saddr));
 	saddr.sun_family = AF_UNIX;
-	memcpy(saddr.sun_path,unix_socket,strlen(unix_socket)+1);
-	
+	if (strlen(unix_socket) > sizeof(saddr.sun_path)) {
+		lpLogger->Log(EC_LOGLEVEL_ERROR, "UNIX domain socket path \"%s\" is too long", unix_socket);
+		return -1;
+	}
+	strncpy(saddr.sun_path, unix_socket, sizeof(saddr.sun_path));
 	unlink(unix_socket);
 
 	if (bind(s, (struct sockaddr*)&saddr, 2 + strlen(unix_socket)) == -1) {
